refactor: MVC패턴을 사용하기 위해 App클래스에 있던 로직 Controller 클래스로 이동
